requirements.txt
azure-ai-projects==1.0.0
azure-identity
openai
python-dotenv
gradio
